backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0
scikit-learn>=1.0.2
scikit-surprise>=1.1.3
numpy>=1.21.0
pandas>=1.3.0

# Frontend Dependencies
react-intersection-observer          # For infinite scroll
recharts                            # For visualization of recommendation stats
heroicons/react                    # For icons
framer-motion                       # For animations
date-fns                            # For date manipulation
react-hot-toast                     # For notifications
tailwind-merge                      # For combining Tailwind classes
react-use                          # Useful React hooks